# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rural Fitness Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Trainer {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  location: string;
  specialty: string;
  /** @nullable */
  avatarUrl?: string | null;
  certificationLevel: string;
  rating: number;
  totalFollowers: number;
  createdAt: string;
}

export interface TrainerInput {
  name: string;
  bio?: string;
  location: string;
  specialty: string;
  avatarUrl?: string;
  certificationLevel: string;
}

export interface TrainerStats {
  trainerId: number;
  totalRoutines: number;
  totalDietPlans: number;
  totalFollowers: number;
  avgRating: number;
}

export interface Routine {
  id: number;
  trainerId: number;
  /** @nullable */
  trainerName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  difficulty: string;
  durationMinutes: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  isFeatured: boolean;
  totalEnrollments: number;
  createdAt: string;
}

export interface RoutineInput {
  trainerId: number;
  title: string;
  description?: string;
  category: string;
  difficulty: string;
  durationMinutes: number;
  thumbnailUrl?: string;
  isFeatured?: boolean;
}

export interface RoutineUpdate {
  title?: string;
  description?: string;
  category?: string;
  difficulty?: string;
  durationMinutes?: number;
  thumbnailUrl?: string;
  isFeatured?: boolean;
}

export interface DietPlan {
  id: number;
  trainerId: number;
  /** @nullable */
  trainerName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  goal: string;
  dailyCalories: number;
  mealsPerDay: number;
  durationWeeks: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  createdAt: string;
}

export interface DietPlanInput {
  trainerId: number;
  title: string;
  description?: string;
  goal: string;
  dailyCalories: number;
  mealsPerDay: number;
  durationWeeks: number;
  thumbnailUrl?: string;
}

export interface ProgressEntry {
  id: number;
  userId: number;
  routineId: number;
  /** @nullable */
  routineTitle?: string | null;
  completedAt: string;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  caloriesBurned?: number | null;
}

export interface ProgressInput {
  userId: number;
  routineId: number;
  durationMinutes: number;
  notes?: string;
  caloriesBurned?: number;
}

export interface ProgressSummary {
  totalSessions: number;
  totalMinutes: number;
  currentStreak: number;
  topRoutines: Routine[];
}

export interface DashboardSummary {
  totalTrainers: number;
  totalRoutines: number;
  totalDietPlans: number;
  totalProgressSessions: number;
  villagesReached: number;
  featuredRoutines: Routine[];
  topTrainers: Trainer[];
}

export type ListTrainersParams = {
specialty?: string;
search?: string;
};

export type ListRoutinesParams = {
category?: string;
difficulty?: string;
trainerId?: number;
};

export type ListDietPlansParams = {
goal?: string;
trainerId?: number;
};

export type ListProgressParams = {
userId?: number;
routineId?: number;
};