# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rural Fitness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Routine } from './routine';
import type { Trainer } from './trainer';

export interface DashboardSummary {
  totalTrainers: number;
  totalRoutines: number;
  totalDietPlans: number;
  totalProgressSessions: number;
  villagesReached: number;
  featuredRoutines: Routine[];
  topTrainers: Trainer[];
}