# lib/api-zod/src/generated/types/dietPlanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rural Fitness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DietPlanInput {
  trainerId: number;
  title: string;
  description?: string;
  goal: string;
  dailyCalories: number;
  mealsPerDay: number;
  durationWeeks: number;
  thumbnailUrl?: string;
}