# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rural Fitness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  id: number;
  userId: number;
  routineId: number;
  /** @nullable */
  routineTitle?: string | null;
  completedAt: string;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  caloriesBurned?: number | null;
}