# lib/api-zod/src/generated/types/progressInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rural Fitness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressInput {
  userId: number;
  routineId: number;
  durationMinutes: number;
  notes?: string;
  caloriesBurned?: number;
}