# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rural Fitness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Routine } from './routine';

export interface ProgressSummary {
  totalSessions: number;
  totalMinutes: number;
  currentStreak: number;
  topRoutines: Routine[];
}