# lib/api-zod/src/generated/types/routine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rural Fitness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Routine {
  id: number;
  trainerId: number;
  /** @nullable */
  trainerName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  difficulty: string;
  durationMinutes: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  isFeatured: boolean;
  totalEnrollments: number;
  createdAt: string;
}